pyproject.toml
[project]
name = "ancpi-data-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "google-cloud-storage>=3.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "ruff>=0.11.5",
    "pandas>=2.2.3",
    "openpyxl>=3.1.0"

]
package-dir = {"" = "src"}
packages = ["ancpi_aggregator"]